ShopInterface 삭제 PlayerLevelMaxXPData : PlayerState->GameState 이전 GameState GetCostProbability, GetMaxXP 플레이어레벨 패러미터 추가

+++ Source/ProjectPC/Public/GameFramework/PlayerState/PCLevelMaxXPData.h
@@ -16,10 +16,10 @@ struct PROJECTPC_API FPCLevelMaxXPData : public FTableRowBase
 
 public:
 	UPROPERTY(EditAnywhere, BlueprintReadWrite)
-	uint8 PlayerLevel;
+	int32 PlayerLevel;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite)
-	uint8 MaxXP;
+	int32 MaxXP;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite)
 	int32 TotalXP;
