+++ Source/ProjectPC/Public/Shop/PCShopUnitData.h
@@ -20,7 +20,7 @@ public:
 	FName UnitName;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite)
-	uint8 UnitCost;
+	int32 UnitCost;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite)
 	FName UnitClass;
@@ -32,7 +32,7 @@ public:
 	TSoftObjectPtr<UTexture2D> UnitTexture;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite)
-	uint8 UnitCount;
+	int32 UnitCount;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite)
 	FGameplayTagContainer Tag;
